refactor: rename some things
