+++ internal/server/clients/exporter/exporter.go
@@ -1,0 +1,7 @@
+package exporter
+
+import "context"
+
+type Exporter interface {
+	Export(ctx context.Context, events []Record) error
+}
