--- internal/server/clients/file/file.go
@@ -1,11 +1,0 @@
-package file
-
-import (
-	"context"
-
-	"github.com/w-h-a/flags/internal/flags"
-)
-
-type Client interface {
-	Read(ctx context.Context) (map[string]*flags.Flag, error)
-}
