--- internal/server/clients/message/message.go
@@ -1,9 +1,0 @@
-package message
-
-import (
-	"context"
-)
-
-type Client interface {
-	Send(ctx context.Context, diff Diff) error
-}
