--- internal/server/clients/message/options.go
@@ -1,35 +1,0 @@
-package message
-
-import "context"
-
-type Option func(o *Options)
-
-type Options struct {
-	URL     string
-	Secret  string
-	Context context.Context
-}
-
-func WithURL(url string) Option {
-	return func(o *Options) {
-		o.URL = url
-	}
-}
-
-func WithSecret(secret string) Option {
-	return func(o *Options) {
-		o.Secret = secret
-	}
-}
-
-func NewOptions(opts ...Option) Options {
-	options := Options{
-		Context: context.Background(),
-	}
-
-	for _, fn := range opts {
-		fn(&options)
-	}
-
-	return options
-}
