+++ internal/server/clients/notifier/notifier.go
@@ -1,0 +1,9 @@
+package notifier
+
+import (
+	"context"
+)
+
+type Notifier interface {
+	Notify(ctx context.Context, diff Diff) error
+}
