+++ internal/server/clients/notifier/slack/domain.go
