--- internal/server/clients/report/domain.go
@@ -1,10 +1,0 @@
-package report
-
-type Record struct {
-	CreationDate int64  `json:"creationDate"`
-	Key          string `json:"key"`
-	Value        any    `json:"value,omitempty"`
-	Variant      string `json:"variant,omitempty"`
-	Reason       string `json:"reason,omitempty"`
-	ErrorCode    string `json:"errorCode,omitempty"`
-}
