--- internal/server/clients/report/report.go
@@ -1,7 +1,0 @@
-package report
-
-import "context"
-
-type Client interface {
-	Create(ctx context.Context, events []Record) error
-}
